Add typeoffset.cs
Move the partial class definition of TypeOffset from interop.cs to its
own file.

Addresses comment:
https://github.com/pythonnet/pythonnet/pull/958/#discussion_r390686331
