+++ src/runtime/Python.Runtime.csproj
@@ -1,4 +1,4 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
+<?xml version="1.0" encoding="utf-8"?>
 <Project DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003" ToolsVersion="4.0">
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
@@ -151,6 +151,7 @@
     <Compile Include="runtime_state.cs" />
     <Compile Include="typemanager.cs" />
     <Compile Include="typemethod.cs" />
+    <Compile Include="typeoffset.cs" />
     <Compile Include="Util.cs" />
     <Compile Include="platform\Types.cs" />
     <Compile Include="platform\LibraryLoader.cs" />
